+++ packages/sdk/src/models/table.ts
@@ -1584,7 +1584,7 @@ class Table extends AbstractModel<TableData, WatchableTableKey> {
      *
      * Checks whether the current user has permission to create a field in this table.
      *
-     * Accepts partial input, in the same format as {@link unstable_createFieldAsync}.
+     * Accepts partial input, in the same format as {@link createFieldAsync}.
      *
      * Returns `{hasPermission: true}` if the current user can update the specified record,
      * `{hasPermission: false, reasonDisplayString: string}` otherwise. `reasonDisplayString` may be
@@ -1596,14 +1596,14 @@ class Table extends AbstractModel<TableData, WatchableTableKey> {
      *
      * @example
      * ```js
-     * const createFieldCheckResult = table.unstable_checkPermissionsForCreateField();
+     * const createFieldCheckResult = table.checkPermissionsForCreateField();
      *
      * if (!createFieldCheckResult.hasPermission) {
      *     alert(createFieldCheckResult.reasonDisplayString);
      * }
      * ```
      */
-    unstable_checkPermissionsForCreateField(
+    checkPermissionsForCreateField(
         name?: string,
         type?: FieldType,
         options?: FieldOptions | null,
@@ -1628,11 +1628,11 @@ class Table extends AbstractModel<TableData, WatchableTableKey> {
     /**
      * _Beta feature with unstable API. May have breaking changes before release._
      *
-     * An alias for `unstable_checkPermissionsForCreateField(name, type, options).hasPermission`.
+     * An alias for `checkPermissionsForCreateField(name, type, options).hasPermission`.
      *
      * Checks whether the current user has permission to create a field in this table.
      *
-     * Accepts partial input, in the same format as {@link unstable_createFieldAsync}.
+     * Accepts partial input, in the same format as {@link createFieldAsync}.
      *
      * @param name name for the field. must be case-insensitive unique for the table
      * @param type type for the field
@@ -1640,19 +1640,19 @@ class Table extends AbstractModel<TableData, WatchableTableKey> {
      *
      * @example
      * ```js
-     * const canCreateField = table.unstable_hasPermissionToCreateField();
+     * const canCreateField = table.hasPermissionToCreateField();
      *
      * if (!canCreateField) {
      *     alert('not allowed!');
      * }
      * ```
      */
-    unstable_hasPermissionToCreateField(
+    hasPermissionToCreateField(
         name?: string,
         type?: FieldType,
         options?: FieldOptions | null,
     ): boolean {
-        return this.unstable_checkPermissionsForCreateField(name, type, options).hasPermission;
+        return this.checkPermissionsForCreateField(name, type, options).hasPermission;
     }
 
     /**
@@ -1680,8 +1680,8 @@ class Table extends AbstractModel<TableData, WatchableTableKey> {
      * @example
      * ```js
      * async function createNewSingleLineTextField(table, name) {
-     *     if (table.unstable_hasPermissionToCreateField(name, FieldType.SINGLE_LINE_TEXT)) {
-     *         await table.unstable_createFieldAsync(name, FieldType.SINGLE_LINE_TEXT);
+     *     if (table.hasPermissionToCreateField(name, FieldType.SINGLE_LINE_TEXT)) {
+     *         await table.createFieldAsync(name, FieldType.SINGLE_LINE_TEXT);
      *     }
      * }
      *
@@ -1690,8 +1690,8 @@ class Table extends AbstractModel<TableData, WatchableTableKey> {
      *         icon: 'check',
      *         color: 'greenBright',
      *     };
-     *     if (table.unstable_hasPermissionToCreateField(name, FieldType.CHECKBOX, options)) {
-     *         await table.unstable_createFieldAsync(name, FieldType.CHECKBOX, options);
+     *     if (table.hasPermissionToCreateField(name, FieldType.CHECKBOX, options)) {
+     *         await table.createFieldAsync(name, FieldType.CHECKBOX, options);
      *     }
      * }
      *
@@ -1701,13 +1701,13 @@ class Table extends AbstractModel<TableData, WatchableTableKey> {
      *             name: 'iso',
      *         },
      *     };
-     *     if (table.unstable_hasPermissionToCreateField(name, FieldType.DATE, options)) {
-     *         await table.unstable_createFieldAsync(name, FieldType.DATE, options);
+     *     if (table.hasPermissionToCreateField(name, FieldType.DATE, options)) {
+     *         await table.createFieldAsync(name, FieldType.DATE, options);
      *     }
      * }
      * ```
      */
-    async unstable_createFieldAsync(
+    async createFieldAsync(
         name: string,
         type: FieldType,
         options?: FieldOptions | null,